bf: add a special treatment for the MD5 of MPUs
Change of the MD5 regexp to also accept "md5-numparts" structure of MPU.
If the MD5 contains dash (-) then it is an MPU MD5 so add a special
processing to transmit the numparts in a specific header.
Cloudserver shall reconstitute the original MD5 based on those 2 headers
for the MPU objects.